requirements.txt
# Core (pin loosely; user may have newer globally)
python-dotenv>=1.0.0
huggingface-hub>=0.25.0
requests>=2.32.0
numpy>=1.26.0
pandas>=2.0.0
duckdb>=1.0.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
torch>=2.0.0
rank-bm25>=0.2.2

# LLM & agents (keep 0.3.x line together; loose pins cause pip to backtrack for minutes)
langchain>=0.3.28,<0.4
langchain-core>=0.3.83,<0.4
langchain-groq>=0.2.0
langchain-google-genai>=2.1.12,<3
langgraph>=0.2.0

# UI
gradio>=5.0.0

# Utils
tenacity>=8.0.0
tqdm>=4.66.0
pyarrow>=14.0.0

# Observability (same minor line avoids resolver fighting grpc/protobuf stacks)
opentelemetry-api>=1.34.0,<1.35
opentelemetry-sdk>=1.34.0,<1.35
opentelemetry-exporter-otlp>=1.34.0,<1.35

# Dev / CI
pytest>=8.0.0
